+++ src/include/sof/lib/alloc.h
@@ -10,20 +10,12 @@
 #define __SOF_LIB_ALLOC_H__
 
 #include <sof/bit.h>
-#include <sof/common.h>
-#include <sof/lib/cache.h>
-#include <sof/lib/memory.h>
-#include <sof/sof.h>
-#include <sof/spinlock.h>
 #include <sof/string.h>
 #include <sof/trace/trace.h>
 #include <user/trace.h>
 #include <config.h>
 #include <stddef.h>
 #include <stdint.h>
-
-struct dma_copy;
-struct dma_sg_config;
 
 #define trace_mem_error(__e, ...) \
 	trace_error(TRACE_CLASS_MEM, __e, ##__VA_ARGS__)
@@ -59,55 +51,6 @@ enum mem_zone {
 /* heap zone flags */
 #define SOF_MEM_FLAG_SHARED	BIT(0)
 
-struct mm_info {
-	uint32_t used;
-	uint32_t free;
-};
-
-struct block_hdr {
-	uint16_t size;		/* size in blocks for continuous allocation */
-	uint16_t used;		/* usage flags for page */
-	void *unaligned_ptr;	/* align ptr */
-} __packed;
-
-struct block_map {
-	uint16_t block_size;	/* size of block in bytes */
-	uint16_t count;		/* number of blocks in map */
-	uint16_t free_count;	/* number of free blocks */
-	uint16_t first_free;	/* index of first free block */
-	struct block_hdr *block;	/* base block header */
-	uint32_t base;		/* base address of space */
-};
-
-#define BLOCK_DEF(sz, cnt, hdr) \
-	{.block_size = sz, .count = cnt, .free_count = cnt, .block = hdr, \
-	 .first_free = 0}
-
-struct mm_heap {
-	uint32_t blocks;
-	struct block_map *map;
-	uint32_t heap;
-	uint32_t size;
-	uint32_t caps;
-	struct mm_info info;
-};
-
-/* heap block memory map */
-struct mm {
-	/* system heap - used during init cannot be freed */
-	struct mm_heap system[PLATFORM_HEAP_SYSTEM];
-	/* system runtime heap - used for runtime system components */
-	struct mm_heap system_runtime[PLATFORM_HEAP_SYSTEM_RUNTIME];
-	/* general heap for components */
-	struct mm_heap runtime[PLATFORM_HEAP_RUNTIME];
-	/* general component buffer heap */
-	struct mm_heap buffer[PLATFORM_HEAP_BUFFER];
-
-	struct mm_info total;
-	uint32_t heap_trace_updated;	/* updates that can be presented */
-	spinlock_t lock;	/* all allocs and frees are atomic */
-};
-
 /* heap allocation and free */
 void *_malloc(enum mem_zone zone, uint32_t flags, uint32_t caps, size_t bytes);
 void *_zalloc(enum mem_zone zone, uint32_t flags, uint32_t caps, size_t bytes);
@@ -119,8 +62,6 @@ void *_brealloc(void *ptr, uint32_t flags, uint32_t caps, size_t bytes,
 void rfree(void *ptr);
 
 #if CONFIG_DEBUG_HEAP
-
-#include <sof/trace/trace.h>
 
 #define rmalloc(zone, flags, caps, bytes)				\
 	({void *_ptr;							\
@@ -239,26 +180,5 @@ void *rzalloc_core_sys(int core, size_t bytes);
 
 int rstrlen(const char *s);
 int rstrcmp(const char *s1, const char *s2);
-
-/* Heap save/restore contents and context for PM D0/D3 events */
-uint32_t mm_pm_context_size(void);
-int mm_pm_context_save(struct dma_copy *dc, struct dma_sg_config *sg);
-int mm_pm_context_restore(struct dma_copy *dc, struct dma_sg_config *sg);
-
-/* heap initialisation */
-void init_heap(struct sof *sof);
-
-/* frees entire heap (supported for slave core system heap atm) */
-void free_heap(enum mem_zone zone);
-
-/* status */
-void heap_trace_all(int force);
-void heap_trace(struct mm_heap *heap, int size);
-
-/* retrieve memory map pointer */
-static inline struct mm *memmap_get(void)
-{
-	return sof_get()->memory_map;
-}
 
 #endif /* __SOF_LIB_ALLOC_H__ */
