+++ src/init/init.c
@@ -12,9 +12,9 @@
 #include <sof/debug/panic.h>
 #include <sof/drivers/interrupt.h>
 #include <sof/init.h>
-#include <sof/lib/alloc.h>
 #include <sof/lib/cpu.h>
 #include <sof/lib/memory.h>
+#include <sof/lib/mm_heap.h>
 #include <sof/lib/notifier.h>
 #include <sof/lib/pm_runtime.h>
 #include <sof/platform.h>
