+++ src/platform/imx8/lib/memory.c
@@ -5,7 +5,7 @@
 // Author: Daniel Baluta <daniel.baluta@nxp.com>
 
 #include <sof/common.h>
-#include <sof/lib/alloc.h>
+#include <sof/lib/mm_heap.h>
 #include <sof/lib/memory.h>
 #include <sof/platform.h>
 #include <sof/sof.h>
