+++ test/cmocka/src/lib/alloc/alloc.c
@@ -13,6 +13,7 @@
 
 #include <sof/sof.h>
 #include <sof/lib/alloc.h>
+#include <sof/lib/mm_heap.h>
 #include <ipc/header.h>
 #include <ipc/topology.h>
