+++ tools/testbench/alloc.c
@@ -12,6 +12,7 @@
 #include <stdio.h>
 #include <malloc.h>
 #include <sof/lib/alloc.h>
+#include <sof/lib/mm_heap.h>
 #include "testbench/common_test.h"
 
 /* testbench mem alloc definition */
